\name{anRpackage-package}
\alias{anRpackage-package}
\alias{anRpackage}
\docType{package}
\title{
Bootstrap Methods for Various Network Estimation Routines
}
\description{
Bootstrap standard errors on various network estimation routines, such as EBICglasso from the qgraph package and IsingFit from the IsingFit package. See \code{\link{bootnet}}
}
\details{
\tabular{ll}{
Package: \tab bootnet\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-03-01\cr
License: \tab GPL-2\cr
}

}
\author{
Sacha Epskamp

Maintainer: Sacha Epskamp <mail@sachaepskamp.com>
}
\seealso{
\code{\link{bootnet}}
}
