\name{null}
\alias{null}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Returns NULL
}
\description{
This function simply returns \code{NULL}.
}
\usage{
null(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
Anything
}
}
\author{
Sacha Epskamp <mail@sachaepskamp.com>
}

\examples{
null("Not NULL")
}
