\name{print.bootnet}
\alias{print.bootnet}
\alias{print.bootnetResult}
\alias{summary.bootnetResult}
\title{
Print method for bootnet and bootnetResult objects
}
\description{
Prints a short overview of the results of \code{\link{bootnet}}
}
\usage{
\method{print}{bootnet}(x, ...)
\method{print}{bootnetResult}(x, ...)
\method{summary}{bootnetResult}(object, ...)
}
\arguments{
  \item{x}{
A \code{bootnet} or \code{bootnetResult} object
}
  \item{object}{
A \code{bootnetResult} object
}
  \item{\dots}{
Not used.
}
}
\author{
Sacha Epskamp <mail@sachaepskamp.com>
}