\name{bootnet-package}
 \alias{bootnet-package}
\docType{package}
\title{
Bootstrap Methods for Various Network Estimation Routines
}
\description{
Bootstrap standard errors on various network estimation routines, such as EBICglasso from the qgraph package and IsingFit from the IsingFit package. See \code{\link{bootnet}}
}

\author{
Sacha Epskamp

Maintainer: Sacha Epskamp <mail@sachaepskamp.com>
}
\seealso{
\code{\link{bootnet}}
}
